adjust token card derivation
